requirements.txt
apache-airflow-providers-google
apache-airflow-providers-hashicorp
# apache-airflow-providers-microsoft-azure